+++ packages/components/src/Actions/Actions.component.js
@@ -1,8 +1,9 @@
 import PropTypes from 'prop-types';
 import React from 'react';
-import { ButtonGroup, OverlayTrigger } from 'react-bootstrap';
+import { ButtonGroup } from 'react-bootstrap';
 import classNames from 'classnames';
 import Action from './Action';
+import OverlayTrigger from '../OverlayTrigger';
 import Inject from '../Inject';
 
 function getButtonGroupProps(props) {