+++ packages/design-system/src/components/Layout/docs/Layout.stories.tsx
@@ -6,8 +6,8 @@ import Column from '../Column';
 
 export default {
 	title: 'Components/Layout',
-
 	parameters: {
+		status: { figma: 'wip', storybook: 'ok', react: 'ok', i18n: 'ok' },
 		component: Layout,
 	},
 };