+++ packages/design-system/src/components/ThemeProvider/ThemeProvider.tsx
@@ -15,6 +15,10 @@ const ThemeProvider = ({ theme = light, children }: ThemeProviderProps<any>) => 
 	const context = useContext(ThemeContext);
 
 	React.useEffect(() => {
+		document.body.dataset.theme = selectedTheme === light ? 'light' : 'dark';
+	}, [selectedTheme]);
+
+	React.useEffect(() => {
 		setSelectedTheme(theme);
 	}, [theme]);